Execution of assignment rules

+++ ProcessMaker/Traits/ProcessTaskAssignmentsTrait.php
@@ -26,6 +26,9 @@ trait ProcessTaskAssignmentsTrait
 
     public static function updateTaskAssignments(Process $process)
     {
+        if (!$process->exists) {
+            return;
+        }
         $process->assignments()->delete();
         $definitions = $process->getDefinitions(true);
         if ($definitions) {
